Implement peer refresh functionality into the peers.js script
This can be invoked by starting the script with "peers.js refresh". The
operation will loop through all the peers in the database, double checking
the geoip location and making sure that it is correct.

As IPAPI now seem to be ignoring requests without a user agent field, a
"iquidus/explorer" user agent header is now added to the request.